Nioclas Regnault: -fix a remaining integer overflow in FQHETorusParticleEntanglementSpectrumOperation
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3531 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Architecture/ArchitectureOperation/FQHETorusParticleEntanglementSpectrumOperation.cc
@@ -388,7 +388,7 @@ bool FQHETorusParticleEntanglementSpectrumOperation::ArchitectureDependentApplyO
 	  architecture->ReceiveFromSlave(i, &TmpValues, TmpNbrValues);
 	  this->NbrNonZeroElements += TmpValues;
 	}
-      TmpValues = (int) this->NbrNonZeroElements;      
+      TmpValues = this->NbrNonZeroElements;      
       architecture->BroadcastToSlaves(&TmpValues, TmpNbrValues);
     }
   else
